modified:   get_rvec_tvec.cpp 	modified:   include/utility.h 	modified:   laser_dot.cpp 	modified:   CMakeFiles/Makefile.cmake 	modified:   CMakeFiles/Makefile2 	modified:   CMakeFiles/TargetDirectories.txt 	modified:   CMakeFiles/laser_dot.dir/CXX.includecache 	modified:   CMakeFiles/laser_dot.dir/depend.internal 	modified:   CMakeFiles/laser_dot.dir/depend.make 	modified:   CMakeFiles/laser_dot.dir/laser_dot.cpp.o 	modified:   CMakeFiles/planes_intersection.dir/progress.make 	modified:   CMakeFiles/progress.marks 	modified:   CMakeLists.txt 	modified:   Makefile 	modified:   laser_dot 	deleted:    utility/include/utility.h

+++ laser_dot.cpp
@@ -130,11 +130,12 @@ int main(int argc, char* argv[])
 				cam_frame_temp0 = cv::Mat(ptrGrabResult0->GetHeight(), ptrGrabResult0->GetWidth(), CV_8UC3, (uint8_t *) pylonImage0.GetBuffer());
 
 				src = cam_frame_temp0.clone();
+
 				//Test intersection function
-				int N1[3] = {10,8,3};
-				int N2[3] = {2,6,5};
-				int point1[3] = {10,5,5};
-				int point2[3] = {2,2,2};
+				double N1[3] = {10,8,3};
+				double N2[3] = {2,6,5};
+				double point1[3] = {10,5,5};
+				double point2[3] = {2,2,2};
     			//intersection::intersectionLine(N1, N2, point1, point2);
 
 				cv::Point lineStart(400,100);
